Fix functional tests scripts.
