+++ package.json
@@ -17,7 +17,7 @@
     "sandbox": "npm run build:json && reactor-sandbox",
     "build": "./scripts/build.mjs",
     "build:prod": "npm run clean && NODE_ENV=production npm run build",
-    "build:json": "./scripts/buildExtensionManifest.js",
+    "build:json": "./scripts/buildExtensionManifest.mjs",
     "package": "npm run build:prod && npx @adobe/reactor-packager",
     "dev": "./scripts/dev.mjs",
     "lint": "eslint \"*.{js,jsx}\" \"{src,test,scripts}/**/*.{js,jsx}\" --cache --fix",
@@ -27,13 +27,13 @@
     "test:unit:coverage": "karma start --single-run --reporters spec,coverage",
     "test:unit:watch": "karma start",
     "test:unit:watch:chrome": "karma start --browsers Chrome",
-    "test:functional": "./scripts/functional.js",
-    "test:functional:saucelabs:firefox": "./scripts/functional.js --firefox",
-    "test:functional:saucelabs:chrome": "./scripts/functional.js --chrome",
-    "test:functional:saucelabs:safari": "./scripts/functional.js --safari",
-    "test:functional:saucelabs:edge": "./scripts/functional.js --edge",
-    "test:functional:watch": "./scripts/functional.js --watch",
-    "test:functional:watch:debug": "echo \"open chrome://inspect in Chrome and click 'inspect'\" && node --inspect-brk ./scripts/functional.js --watch",
+    "test:functional": "./scripts/functional.mjs",
+    "test:functional:saucelabs:firefox": "./scripts/functional.mjs --firefox",
+    "test:functional:saucelabs:chrome": "./scripts/functional.mjs --chrome",
+    "test:functional:saucelabs:safari": "./scripts/functional.mjs --safari",
+    "test:functional:saucelabs:edge": "./scripts/functional.mjs --edge",
+    "test:functional:watch": "./scripts/functional.mjs --watch",
+    "test:functional:watch:debug": "echo \"open chrome://inspect in Chrome and click 'inspect'\" && node --inspect-brk ./scripts/functional.mjs --watch",
     "test": "npm-run-all --parallel test:unit test:functional",
     "token": "node scripts/token.js",
     "version": "./scripts/version.sh"
