+++ scripts/buildExtensionManifest.mjs
@@ -10,16 +10,15 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTA
 OF ANY KIND, either express or implied. See the License for the specific language
 governing permissions and limitations under the License.
 */
-const buildExtensionManifest = require("./helpers/buildExtensionManifest");
+import buildExtensionManifest from "./helpers/buildExtensionManifest.mjs";
 
-buildExtensionManifest()
-  .then(resultPath => {
-    // eslint-disable-next-line no-console
-    console.log(
-      "\x1b[32m%s\x1b[0m",
-      `✅ Extension manifest written to ${resultPath}`
-    );
-  })
-  .catch(e => {
-    console.error(e);
-  });
+try {
+  const resultPath = await buildExtensionManifest();
+  // eslint-disable-next-line no-console
+  console.log(
+    "\x1b[32m%s\x1b[0m",
+    `✅ Extension manifest written to ${resultPath}`
+  );
+} catch (e) {
+  console.error(e);
+}
